// apps/server/functions/src/app/graphql.schema.interface.ts
export interface OriginLocationInput {
    city: string;
    state: string;
    zipcode?: Nullable<string>;
}
// ...
export interface Truckstop {
    __typename?: 'Truckstop';
    isLoggedIn?: Nullable<boolean>;
    startNewSearch?: Nullable<boolean>;
    operationMsg?: Nullable<string>;
    isSearchContainerOpen?: Nullable<boolean>;
    originLocationInputState?: Nullable<OriginLocationInputState>;
    loadsContainerTabsState?: Nullable<LoadsContainerTabsState>;
    loadSearch?: Nullable<LoadSearch>;
}

export interface OriginLocationInputState {
    __typename?: 'OriginLocationInputState';
    originInput: string;
    inputValue: string;
    status: string;
}

export interface LoadsContainerTabsState {
    __typename?: 'LoadsContainerTabsState';
    openTabsCityStates: TabCityState[];
    activeTabIndex: number;
}

export interface TabCityState {
    __typename?: 'TabCityState';
    originLocation: string;
    destinationLocation: string;
}

export interface LoadSearch {
    __typename?: 'LoadSearch';
    action: string;
    city?: Nullable<string>;
    state: string;
}
// ...
type Nullable<T> = T | null;